MAX_COST_FOR_CONSTANT_PROPAGATION = 1000
Summary:
We only simulate constant-propagation and local-dce for callees with constant-arguments with the callee doesn't exceed a certain size, in order to keep the Redex runtime overhead down.

Given the additional use of the analysis result (to infer no-return callees), and the reduction in inlining overhead (by subtracting dead blocks), we can bump up the limit for additional gains without noticable overhead.

Reviewed By: kuganv

Differential Revision: D29074262

fbshipit-source-id: 92cbe6e915e7ddd08d0bd372b758e3d906e499a9