cut off control-flow after no-return callsite
Summary:
When a method such as
```
   void checkMe(object arg) {
     if (arg == null) throw new FancyException("SuperBla");
   }
```

is invoked for particular "constant" arguments such as
```
   checkMe(null);
   moreSuperLongCode();
```

then, for the given constant arguments, `checkMe` does not have a feasible control-flow path that ends with a normal `return` statement. Thus, the subsequent normal (non-exceptional) control-flow at the callsite (`moreSuperLongCode`) is dead.

In fact, the inliner already does some relevant analysis in this direction: For all callees (that are inlining candidates) with constant arguments, it simulates constant-prop and local-dce to determine the size of the eventually remaining instructions after applying those optimizations in earnest. This is used as part of a cost function to determine if inlining any particular call is beneficial in terms of size (and thus most likely performance).

As it turns out, it is rather trivial --- both in terms of added runtime complexity in Redex, and amount of code --- to extent this functionality to also determine if for a particular callsite with constant arguments the callee does not have a normal return path.

And then, the inliner will now (optionally) insert a (unreachable)
```
  throw null;
```
statement after such a callsite, to cut off the remaining control-flow. This is generally a good idea, and it also helps the inliner to inline more given the reduced code size. We call `remove_unreachable_blocks` to get a precise "estimate" for the resulting instruction size.

Bonus:
- I refactored `remove_unreachable_blocks` a bit to avoid that it calls `recompute_registers_size`, as the inliner manages the registers size separately. Instead, there's an extra return component that signals that this might be beneficial, and it's up to each caller to do it or not. As it turns out, there were some callers (LocalDcePass, ThrowPropagationPass), that would run another unconditional `recompute_registers_size` anyway, and there we can now save a run.
- `remove_unreachable_blocks` frees blocks, and that includes freeing the instructions in them (!). Thus, the instruction-pointers could get reused for something else, and it is no longer safe to use disappeared pointers for any purpose. Thus, I added some code to track well-defined `remaining_callsites`. This is also employed after intermediate-shrinking, effectively fixing a potential problem when that feature is enabled.

Reviewed By: beicy

Differential Revision: D29014533

fbshipit-source-id: ba0917a4044474cb4e98f1b1e9aa224714a43a8c