Expand proguard parsing for multilingual and dump (#618)
Summary:
`-dump` is an option in proguard that we use but which wasn't supported by the parser and caused the proguard config to be rejected. So, we added support for dump into the parser.

In addition, we use multi-lingual external dependencies that have fields written in non-ascii text that, by Java spec, is legal but which isn't allowed by the proguard parser. Since writing an accurate port of Java's `Character.isJavaIdenfier*()` family of methods is non-trivial and error-prone, we instead took the path of checking against a small selection of illegal characters and then allowing anything isn't explicitly disallowed. This will allow some false-positives, but will result in the acceptance of multi-lingual proguard configurations.

Pull Request resolved: https://github.com/facebook/redex/pull/618

Reviewed By: agampe

Differential Revision: D29274863

Pulled By: yuxuanchen1997

fbshipit-source-id: 59adf43f2e70c147e874427d1e5bd2502872d1bf