Don't outline from methods with monitor instructions
Summary:
The outliner may outline a code fragment that has no throw-edges attached to it, but by inserting a `invoke-static`, even that it may not throw in practice, the Android verifier demands a catch-all handler if the code appears in a nested `monitor` setting.

To avoid an Android verifier failure, we disable outlining from methods with monitor instructions for the time being.

Reviewed By: wsanville

Differential Revision: D29309925

fbshipit-source-id: 15dd0ff6d19791c05e20fb88c89bd35d560b9f58