Default `m_always_inject = true`
Summary:
As title.

This is better for tracking down optimization issues, and detangle profile ambiguities. Without this, a missing `SourceBlock` may mean "an optimization forgot to update this, but it could be hot" or "there was no profile, so the block is cold."

There is a corresponding wall time overhead. This has been deemed acceptable at this ppoint.

Reviewed By: NTillmann

Differential Revision: D28967350

fbshipit-source-id: d07a01102573b5e6310df822875ab2c62b4edadd