Reverse Promote D29309925: Don't outline from methods with monitor instructions
Summary:
The outliner may outline a code fragment that has no throw-edges attached to it, but by inserting a invoke-static, even that it may not throw in practice, the Android verifier demands a catch-all handler if the code appears in a nested monitor setting.
To avoid an Android verifier failure, we disable outlining from methods with monitor instructions for the time being.

Reviewed By: agampe

Differential Revision: D29318827

fbshipit-source-id: cdc569dcec726428e06a579661527e0d5252dce7