Special handling well known external java.lang.* classes
Summary:
IR Typechecker does not check type assigning when external classes are involved.
This improvement assumes that the hierarchies of the well known java.lang.* are stable across Android versions.
When we have the class definitions of these external java.lang.* classes, we can perform the regular type inheritance check.

Reviewed By: thezhangwei

Differential Revision: D29305779

fbshipit-source-id: a0482a4ab83f6da4574212b8524264e2d7b17c24