DexTypeValue join converges faster with widen_with
Summary:
When `widen_with` is invoked by the Fixed Point Iterator, the domain converges faster than the `join_with` operation.

The rationale is that for certain obscure input, the dex type of a value can flip flop in a loop causing the fixed point iterator to run indefinitely. The old `widen_with` simply defer to the existing `join_with`. This change makes it converge faster to avoid infinite iteration.

Reviewed By: ssj933

Differential Revision: D29541977

fbshipit-source-id: ee022a9b935869d8d8483675f139407c1e0bc30f