do not consider outlined code itself when computing result type of const results
Summary:
When a fragment of code produces a set of constants, we look for possible uses to determine what the formal result type of an outlined methods should be.

After the point of convergence (the end of the outlined code fragment), the strictest use of the register can be used to determine the formal type. To this end, uses WITHIN the outlined code fragment should NOT be considered. This is getting addressed here.

Reviewed By: agampe

Differential Revision: D29530941

fbshipit-source-id: 25d562bd0138f6a479bb49541a90b9625866414a