Make redex/service part of the public redex installation
Summary:
I recently landed D24062337, which adds redex global type analysis to mariana trench.

However, this has broken our open source cmake build because it imports from GlobalTypeAnalysis.h. This header file is in redex/service rather than redex/libredex and I don't think it's available as part of the public headers for redex (as far as I can tell looking at fbsource/fbandroid/native/redex/CMakeLists.txt). Should we make this public, or is there another API I can include instead?

Chatted with yuxuanchen1997 and it sounds like it's OK to make service public, so tentatively trying this out in a diff for now. I'm not familiar with cmake at all, so let me know if I'm barking up the wrong tree.

Note: redex is expected to be installed independently as a dependency to open source MT builds. It's found with D26116776.

Reviewed By: arthaud

Differential Revision: D29668450

fbshipit-source-id: c05e71f50be451129e3143f209cf2784fe575baa