Inline within fully disconnected self-recursive methods
Summary:
The "call graph" the inliner operates on is not complete, as we only consider potentially inlinable callees.

As it turns out, some potentially inlinable callees are not reachable from any caller that is not a callee.

This is getting fixed here, enabling a bit more inlining.

(While not very impactful on its own, it makes the next diff more straightforward.)

Reviewed By: beicy

Differential Revision: D29412088

fbshipit-source-id: f2045ae0333a14867b1022f3169db7f4846f4db7