caller_callee, callee_caller as counted map
Summary:
Instead of representing the range of those datastructure as confusing vectors with duplicates, we here switch them to unordered maps with the number of occurrences in the range. This better reflects intent, and may be a bit faster, we avoid a few linear operations.

This is a behavior-preserving change.

Reviewed By: kuganv

Differential Revision: D29415412

fbshipit-source-id: 3c33672181d672763d2a39e96074ea61748ca759