parallelize and "enhance" true-virtual method filtering
Summary:
We currently don't insert casts when inlining true virtuals, thus we cannot accommodate true virtual methods that in any form use the receiver. This is getting implemented here, slightly the earlier approach of looking for any type/field/methods in the code, whether operating on the receiver or not.

Also, the computation is parallelized.

Reviewed By: beicy

Differential Revision: D29396001

fbshipit-source-id: 8b036b76d9b617596a3e0f8af59bacdc547216df