Inline more intra-dex
Summary:
The way the IntraDex inlining was originally written was a bit limiting: It rules out any (caller, callee) pair if the callee has **any** caller in a different dex. That's overly limiting. In this diff, inlining is enabled for any particular (caller, callee), as long as the particular caller and callee are within the same dex. Still, additional inlining at any particular call-site will only happen if it's predicted to not increase overall code size.

For the time being, only the PerfMethodInlinePass retains the old limiting behavior (see comment in code), as otherwise it would inline way more call-sites than it used to, leading to huge code size increases.

Future work: We could inline even more if we more closely inspect the callee, and see if it contains any cross-dex references, in general, and for each call-site after pruning via constant-prop and local-dce.

Reviewed By: agampe

Differential Revision: D29624444

fbshipit-source-id: 2167c5d14094533622f13ac6a29382d510b3742c