Handle lookup table with kotlin null checking
Summary:
To fix an enum pass crash with recent kotlin upgrade.

The enum-when-map removal tries to match bellow pattern where the switch key is from a lookup table.
```
 * Pattern we are trying to match:
 *  SGET_OBJECT <LookupTable>;.table
 *  MOVE_RESULT_PSEUDO <v_field>
 *  ...
 *  INVOKE_VIRTUAL <v_enum> <Enum>;.ordinal:()
 *  MOVE_RESULT <v_ordinal>
 *  ...
 *  AGET <v_field>, <v_ordinal>
 *  MOVE_RESULT_PSEUDO <v_dest>
 *  ...
 *  *_SWITCH <v_dest>            ; or IF_EQZ <v_dest> <v_some_constant>
```
With the kotlin upgraded, the switch key can be a constant from a enum null checking.
```
 *  IF_NEZ <v_enum> :NON-NULL-LABEL
 *  CONST <v_dest> -1
 *  GOTO :SWITCH-LABEL
 * ...
 * *_SWITCH <v_dest>
```
To handle the case, the diff changes abstract domain of the data flow analysis for the switch key, to allow it be a negative constant value.

Reviewed By: amnn

Differential Revision: D29647995

fbshipit-source-id: 997dfd43589c4aac0fcb07dfa6c20fa53489d7be