Include more true virtual calls
Summary:
Up to now, we only considered true-virtual methods if the callee is known and does not use the receiver at all.
The reason is that invoking a true-virtual method implicitly casts the receiver (the first argument) to the type of the true-virtual method class, and the inliner cannot (yet) make this cast explicit.
In this diff, we identify additional situations where such a cast isn't actually needed, i.e. when all uses of the receiver only demand that the receiver has a type that was already guaranteed at the call-site.

Reviewed By: Feng23

Differential Revision: D29689382

fbshipit-source-id: 84acfc7e862f9731b9f6a5279f00a7e3dec7b042