insert casts for true-virtual methods as needed
Summary:
Until now, the inliner avoided considering true-virtual methods if they used the receiver in a way that would require casts to be inserted.
In situations where the required cast is simply to cast to the (unique) known declaring true-virtual method class, we now insert such a cast.
(What's still excluded here is the case of "same implementations" true virtual methods, where finding an appropriate cast type is a bit more involved.)

Reviewed By: Feng23

Differential Revision: D29706534

fbshipit-source-id: bc0020a4f4ebc879e53d45bb35404a8a65bebaca