precisely track whether all true virtual method call sites are well known
Summary:
This is used in the inliner's cost model to decide whether to inline any particular call-site, or all call-sites in order eliminate the method entirely. Thus, we don't filter "root" methods when building the candidates, and leave it up to the inliner to decide whether to use a candidate, or not.

The inliner may now also directly delete true virtual methods if they have been fully inlined, instead of leaving it to RMU. Notice that the inliner already deleted non-true-virtual methods before.

Reviewed By: Feng23

Differential Revision: D29698629

fbshipit-source-id: 883725b47024562f55407e9bfe088c4bae88a947