Check Android verifier quirk around synchronized throwing instructions without catch-all
Summary:
As the long comment on top of MonitorCount.h explains, we have to deal with a strange Android verifier quirk. So far, we'd mark certain "sketchy" incoming methods at the beginning of a Redex run as "no-optimizations" and "don't inline", which is good enough to avoid that those turn into actual Verifier errors.

However, as we've discovered in recent weeks, that's not sufficient, as Redex can create additional "sketchy" methods. In particular, the outliner and the ReduceBooleanBranches pass do that. And then the inliner turns them into actual Verifier errors.

This diff enhances our own IRTypeChecker to detect code that involves mismatched monitor depths and the particular strange Android verifier quirk. It also slightly refines the MonitorCount implementation to properly undo counter changes on exception, so that we don't keep getting into useless "top" states.

Most importantly, this diff also prevents the inliner from turning a sketchy method into something that will actually trigger the strange Android verifier error.

Here's the pseudo-code example from the new unit test that illustrates what we'll not attempt to inline anymore:

```
class Foo {
  void caller() {
    try {
      sketchyCallee();
    } catch (SomeSpecificType) {
    }
  }
  void sketchyCallee() {
    synchronized(this) {
      goto outside;
      inside:
    }
    return;
    outside:
    Bar.canThrowOutsideTry();
    goto inside;
  }
}
```

The `sketchyCallee` contains some interesting code where some possibly exception-throwing code `Bar.canThrowOutsideTry()` can only run within a `synchronized` context, however, is not actually "protected" by any exception handling code. This kind of code is actually emitted by the Java and Kotlin compiler, where the possibly exception-throwing code is indeed static method calls to some helper methods that won't actually throw. And the Android verifier is okay with it.

The real problem would arise when we try to inline the `sketchyCallee` into a try-catch in the `caller`. Normally, inlining code into a try-catch block means that all possible-throwing inlined code gets additional throw-edges to the outer catch block. However, here, when we would add such edges to the `Bar.canThrowOutsideTry()` block, then that block would effectively be within a try-catch region, but **it would not have a catch-all handler** associated with it, a condition which the [Android Verifier will fail the code for].
(http://androidxref.com/9.0.0_r3/xref/art/runtime/verifier/method_verifier.cc#3553).

Besides Java and Kotlin compiler, Redex can also create "sketchy" code e.g. in the outliner and the ReduceBooleanBranches pass, when mutating a block with instructions that cannot throw and that is within a synchronized region, however, the block itself ends with some kind of (conditional) branch, and thus doesn't have any throw-edges associated with.

So we have to do something in the inliner to avoid creating code that will fail the Android Verifier, and we choose to just not inline when that would lead to the Android Verifier violations.

Reviewed By: agampe

Differential Revision: D29781187

fbshipit-source-id: 20ab6132bbdf78d3b4c86eaa208228912fc5b974