Correct the live interval calculation
Summary:
Add a separate file to calculate live intervals.
In current method, live intervals are calculated based on live-in/live-out sets of each basic block. Because the code within a single block must be linear, a vreg's alive instruction range in a block must begin from either block top or the vreg's first def in this block, and must end at either block end or the vreg's last use (if any) in this block.
The previous live interval calculation in FastRegAlloc pass is defined from first Def to  last Use of a vreg in the ircode. This is incorrect because ircode can be nonlinear, and if there're loops, a vreg defined in loop header may need to be alive until end of this loop, even if the last instruction in this loop is neither a Def nor a Use of this vreg.

Reviewed By: beicy

Differential Revision: D29502134

fbshipit-source-id: 3e5505f09894f1aaf59b3de194fe5a3168cd4e34