Don't leave behind handling exit-block pointers when deleting blocks
Summary:
The "exit block" used to be not well managed by the cfg. It is only ever optionally set, or `nullptr` is a legal way when not initialized. But then when the earlier exit block is removed, the `exit_block` property isn't reset to `nullptr`.
This is getting fixed here, so that we don't have to deal with dangling undefined pointers.

Reviewed By: kuganv

Differential Revision: D29892627

fbshipit-source-id: fc9d860948d519dfa641fd04f62074315f5e45c2