Worked around dead block with -1 key in SwitchMap
Summary:
Kotlin1.5 combines null handling block with enum SwitchMap with -1 as the key. When the null checking is dead but null handling block is existing, we get an assertion error in enum pass.
The diff deletes the block with negative key in the switch when they are dead.

code example:
```
public fun useNonNullEnum(obj: EnumType): Int =
  when (obj) {
    null -> 999 // The obj is non-null, this block is deleted.
    EnumType.A -> 0
  }
```

Note: the current implementation is not able to remove the synthetic enum SwitchMap when there is null checking.
code example:
```
public fun useNullableEnum(obj: EnumType?): Int =
  when (obj) {
    null -> 999 // not dead.
    EnumType.A -> 0
  }
```

Reviewed By: amnn

Differential Revision: D30005159

fbshipit-source-id: 4bed61743a7227fef96cc95b3042febcee764476