fill in case-keys for uneditable cfg
Summary:
For reasons not known to me, the cfg-construction would NOT fill in case keys for the non-editable cfg. This is surprising, and leads to subtle semantic differences when analyzing cfgs by constant-propagation. In this diff, we add in the missing case-keys, but don't use them just yet in the constant-propagation analysis. The new TODOs will get addressed in the next diff up the stack.

This is a behavior preserving change.

Reviewed By: Feng23

Differential Revision: D29473765

fbshipit-source-id: ea0114120b50ba42b474a54f2e6644d1dbeee222