Exclude wrapper methods with trivial exception handlers
Summary:
The Synth pass, one of the earliest optimizations in Redex, is essentially doing very specialized inlining transformations, based on code pattern matching.
While some optimizations are guarded by code having to be `is_synthetic`, and thus be in some limited compiler-generated form, the `trivial_method_wrapper` is not guarded in such a way. It checks if a method just calls another method, and doesn't execute any other instructions, and if so, "inlines" it by rewriting the invocation instruction to the wrapper.

Where this can go wrong is code like this (the new test case):
```
class Delta {
  public void foo() {
    Epsilon.wrapper();
 }
}

class Epsilon {
  public static void wrapper() {
    try {
      target();
    } catch (Throwable t) {

    }
  }

  public static void target() {
    // some code...
    java.lang.Integer.parseInt("CrashMe");
  }
}
```

The IRCode of the Epsilon.wrapper() looks like this:
```
[0x7fbdc000acd0] TRY: TRY_START 0x7fbdb8006aa0
[0x7fbdb8003df0] DEBUG: DBG_SET_PROLOGUE_END
[0x7fbdb80091f0] POSITION: Lcom/facebook/redextest/Epsilon;.wrapper:()V(Alpha.java:91)
[0x7fbdb80069e0] OPCODE: INVOKE_STATIC Lcom/facebook/redextest/Epsilon;.target:()V
[0x7fbdc000e8f0] TRY: TRY_END 0x7fbdb8006aa0
[0x7fbdb8006aa0] CATCH: Ljava/lang/Throwable;
[0x7fbdb8009220] POSITION: Lcom/facebook/redextest/Epsilon;.wrapper:()V(Alpha.java:95)
[0x7fbdb8006a40] OPCODE: RETURN_VOID
```

The try/catch mflow markers are clearly there. But the code in the Synth pass, operating with an instruction iterator, doesn't see them, and what remains is what looks like a wrapper code straight instruction sequence.

This is getting fixed here.

Beyond this short-term fix, we should seriously considering replacing this ad-hoc pattern matching code, or the entire Synth pass, with something based on our inliner, so that...
- Positions are not lost
- Source blocks are not lost
- Cross-store references are respected

Reviewed By: kuganv

Differential Revision: D30167547

fbshipit-source-id: f33fb51a99c4699f5d1914a7d21a9ed3b44ff8e7