fixing undefined behavior
Summary: `std::prev(it)` was undefined ever since the pass was upgraded to be to be based on an editable cfg. I think the replacement may have been just lost in some cases. This is getting fixed here, and all the other changes, "upgrading" the iterator to be a cfg-iterator, are ripple effects for consistency.

Reviewed By: agampe

Differential Revision: D30215084

fbshipit-source-id: 2b91b43391bd415211bf2e9e0011feb09f2b8e45