Check if a class is deletable when changing its superclass
Summary:
The code tries to replace all superclass constructor calls but doesn't analyze if the constructor is invoked on "this" object.
We will deprecate the `handle_interface_as_root`. To temporarily work around the bug, we can add DoNotStrip to classes that can trigger the bug.
This diff checks if a class is deletable before changing its superclass.

Reviewed By: thezhangwei

Differential Revision: D30282362

fbshipit-source-id: 376de585820ca2db89d9a205d6869fcde1c874e3