Fix walk_references_for_resource
Summary:
I think `nodes_to_explore.pop()` was destroying some actual data, resulting in wrong outcomes. Change that to a pointer (and dereference it before popping) to keep things intact.

Second problem was that resource table will hold file paths as-if they are read on a device, at which point the concept of "modules" in a bundle do not exist. This meant that we were attempting to read <unpack dir>/res/layout/activity_main.xml which does not exist in an App Bundle. These file paths needed to consider the module name from which they came.

`leaf_string_values` is kind of a misleading name as a result of the change, which I will likely rename in a separate diff.

Reviewed By: ssj933

Differential Revision: D30230001

fbshipit-source-id: 8b58da136889df831b54e0d26e9f5112677f56df