Only remove interfaces with a limited number of implementers
Summary:
To remove an interface with a large number of implementers we have to construct a deep dispatch method.
It's beneficial if we can remove a large number of interfaces. Which in turn enables more classes to be merged.
But the cost of doing that start to be come more and more visible. By limiting the number of implementers, we can avoid building expensive dispatches.

Reviewed By: Feng23

Differential Revision: D30202329

fbshipit-source-id: 3eb75345f0d4bdd884d226ac0a1c1d659e1295c0