don't do quadratic succs() lookup
Summary:
The cfg happens to store switch case information in a form that isn't indexable by case keys.
The way constant-propagation was finding matches effectively made it quadratic in the number of switch cases (for each switch case edge, iterating over all other switch case edges).

We know that there are large methods mainly consisting of many switch cases, so this is not completely irrelevant.

This diff introduces a simple cache that effectively reduces this quadratic behavior to linear.

This is a behavior preserving change.

Reviewed By: Feng23

Differential Revision: D30229587

fbshipit-source-id: dca5d3e5678c25c549e2b54130c49e361d3c0740