Back out "Attempt to remove methods based on candidate list, not ultimately inlined list"
Summary:
Original commit changeset: 95b33e78a07a

Letting the inliner remove true virtual methods was a bad idea, this is getting backed out. The situation we ran into was like this, in the second IntraDexInlinePass run, however, with B.foo() in a different dex, so it didn't get inlined:
```
interface I { void foo(); }
abstract class A implements I {
  void bar() {
    this.foo(); // This would be a ref to A.bar --- a method that doesn't quite exist
  }
}
class B extends A {
  void foo() {}
}
```
And then the inliner wrongly thought that no more references to I.foo() exist, and deleted it. While the reasoning about true virtual methods in the inliner is good enough to guide inlining decisions, it should not be used to make the final decisions on whether a true virtual method may be removed or not. That recently introduced behavior is being backed out here. Instead, we really leave it to the RMU pass which has much more involved logic including reasoning about resources and reflection uses.

Side notes:
- I am trying to make up for the size loss by running RMU more often two diffs up.
- While after this diff the inliner will still remove non-true virtuals, which is has been doing for quite some time, two diffs up also (optionally) disables that behavior and complete leaves all deletion decisions to the RMU pass.
- The time we save by no longer doing computations in the inliner to determine what methods to delete or not turns out to be roughly the same as it takes to run RMU, so we are not losing anything really, and the code gets less confusing. In fact, RMU will delete more methods, fields, and class, being a small win overall.

Reviewed By: ssj933

Differential Revision: D30324447

fbshipit-source-id: e29a3b2b41cd9bd04420c82908a31508a15c15cc