Only create new MergerType.InterfaceMethod for VirtualScope that has method defs
Summary:
When collecting intf methods for `MergerTypes`, we do so by parsing the `VirtualScopes` on all mergeable types.
We try to match an intf `VirtualScope` with existing `InterfaceMethod`s collected on the same `MergerType`.
However, if there's no match, we create a new `InterfaceMethod` placeholder and insert the new `VirtualScope`.

There's one caveat when doing this. That is we should only create a new `InterfaceMethod` for a `VirtualScope` that at least has one method def and that method def is on a mergeable type. Unfortunately, we did not handle this correctly.

This change includes:
- Add a `has_def` check on the incoming `VirtualScope` and only create a new `InterfaceMethod` instance when passing the check.
- Add more comments to make the logic more readable

Reviewed By: Feng23

Differential Revision: D30399392

fbshipit-source-id: 05674b99f9f140d74b5325ca8bd3ef9056f0eabf