don't do quadratic edge removal in constant-prop transform
Summary:
The way constant-propagation was removing edges was effectively quadratic in the number of switch cases, as for each edge, it would iterate over all remaining successor switch case edges.

We know that there are large methods mainly consisting of many switch cases, so this is not completely irrelevant.

This is getting improved here by providing a new overload to many edges in the cfg at once, doing only a single scan.

**This is a behavior preserving transformation.**

Reviewed By: Feng23

Differential Revision: D30232486

fbshipit-source-id: 964dde6fffc6cf3543712011ea70c346eb69cda2