Introducing the Partial-Application pass
Summary:
This pass identifies commonly used constant arguments that methods are invoked with, and then introduces helper functions that bind those arguments if it seems beneficial to reduce the overall code size by rewriting the call-sites. The new helper methods are placed in the same class as the callee. Their name is stable, including a hash derived from the bound constant arguments.

The most interesting part of this optimization, with likely further tuning potential, is the priority-queue based approach to find beneficial subsets of common constant arguments.

While this is similar in spirit to what the InstructionSequenceOutliner does, a major difference is that this pass specifically targets individual method invocations, and it picks up incoming constant arguments based on our existing constant-propagation analysis, not caring where earlier in the code, or in which order, the constant are defined. And ultimately, it picks a beneficial subset of constant arguments regardless of what order they were defined in. In contrast, the InstructionSequenceOutliner requires precise matches of frequently occurring instruction opcodes sequences (but modulo register names) in order outline any particular call-site.

Here's an example of what the optimization does. Let's say there's a method like this:
```
  void foo(int a, int b, Integer c);
```
And it is invoked 10 times as
```
  foo(10, 20, Integer.valueOf(23));
```
And another 10 times as
```
  foo(13, 20, Integer.valueOf(23));
```
Let's say in neither case would a new helper function be beneficial to reduce size. However, when we trim off the first argument, we are left with 20 times
```
  foo(*, 20, Integer.valueOf(23));
```
And this might be beneficial to transform. Then we introduce a helper function like the following.
```
  foo$pa$xxxx(int a) { foo(a, 20, Integer.valueOf(23)); }
```
And rewrite the call-sites to
```
  foo$pa$xxxx(10);
```
and
```
  foo$pa$xxxx(13);
```
respectively.

Note that there might be also entirely different call-sites, or other clusters of constant arguments, that would get transformed separately. That's why this pass is also different from (IP)CP's insertion of (universally unique) constant arguments.

Various safe-guards are in place:
- We won't introduce helper methods that would contain cross-store or non-min-sdk level references.
- We only transform code with the largest root store id (so not in the primary dex, unless there only is one, and not in other auxiliary stores)
- We won't rewrite code that sits in hot blocks in hot methods, or loops in warm methods (reuses logic from InstructionSequenceOutliner), to avoid any significant runtime overhead from the extra method invocations.

We don't do anything special for symbolication. Thus, the new helper methods will appear in stack traces, but probably won't be confusing, as they have names derived from the final callee, appearing as some trampoline method. The code in the helper methods will never fail (except maybe under the most obscure circumstances such as a stack-overflow), and thus will never be on top of a stack trace, and only the top frame is used for symbolication.

Reviewed By: kuganv

Differential Revision: D30060032

fbshipit-source-id: 650604100e6bcb1b85f8353c5dfa1a64a5465579