Make get_hash_from_values return more sensible values by removing source info
Summary:
ResStringPool bytes in the resources.pb file is really just a string pool of file paths of the build machine that prepared the .aab file. This is a little bit unintuitive since the ResStringPool struct exists in resources.arsc file, and is used on-device, the .pb files use it for a completely different purpose!

`Source` messages throughout the rest of the table are indexing into this. We should be fine to normalize them all to a given value to make comparisons easier throughout all subsequent operations.

Note: In debug .aab builds, this source pool is present, but in release style it seems absent.

Reviewed By: ssj933

Differential Revision: D30439518

fbshipit-source-id: 636827d49f3f7f6d4ef571ec404b02cb044a3bed