properly consider interface methods that can have dynamic implementations
Summary:
The code used to erroneously check the property for the (caller) `method`, not `callee`, which makes no sense. This is getting fixed here.

Also, to be even more conservative, we look back at all overridden methods to see if any of those is defined in an interface and has keep rules.

Reviewed By: ssj933

Differential Revision: D30628118

fbshipit-source-id: 7b9a061bf55435b504aee12632146b0a44f7008b