refactor how methods are made static
Summary:
This removes the ugly "make_static" lists that were explicitly carried through multiple APIs.
Instead of that, we now infer those methods from the visibility-changes computation, with which the make-static lists had the be kept in sync.
We can now run `delayed_invoke_direct_to_static` just after with `delayed_visibility_changes_apply`, as we can now derive the information about make-static methods internally.
The deleter gets simplified and no longer has to worry about methods that we'll want to make static.
Most importantly, we can get rid of the `fix_static_invokes` hack in the inliner that was needed in the context of experiments.

Fixed some test cases that were conceptually dubious before, as they declared inconsistent methods ("private virtual").

This is a behavior-preserving change.

Reviewed By: adicatana

Differential Revision: D30621626

fbshipit-source-id: aacb328e2639c6c77cba08b642f0da2b5cb0b4ef