Remove run_before_interdex option
Summary:
While the ClassSplitting pass was originally designed to run as an InterDex plugin, we later realized that it should always run before InterDex. This ability was hacked in later via the `run_before_interdex` option, which had been hard-coded to be `true` ever since.
This diff gets rid of the option, simplifying the code for the `true` case. In particular, this also makes the `should_not_relocate_methods_of_class` parameter in `InterDexPassPlugin::gather_refs` unnecessary, as it's not used by any other plugin.
So this is effectively a behavior preserving change, actually code simplification.

Reviewed By: Feng23

Differential Revision: D31028520

fbshipit-source-id: 7b97b9381e7acdebafe4a3fc7c54b3b03984669d