always pick worst class after overflowing
Summary:
When starting a new dex file (because the last one got full), don't just start with the previously selected class that would have been the best match for what was already added to the previous dex file.

Instead, use the "worst" class after overflowing. This...
- could lead to more stable results over time, as the "worst" class is more stable than one random overflowed class,
- should lead to somewhat better overall size results.

This wasn't possible in the past due to stateful interdex-plugins, but since those have been cleaned up, we can do the change now.

Reviewed By: Feng23

Differential Revision: D31007869

fbshipit-source-id: bdde55b0926414c219eaea4e43ad44345789caa7