More ref-checking
Summary:
This adds one more layer of ref-checking to the Synth pass, preventing e.g. the inlining of a field access across a store boundary, even if the invocation to the getter method already represented a cross-store access.

This makes the SynthPass a little bit more conservative, inline with what the inliner also already does.

Reviewed By: beicy

Differential Revision: D30610935

fbshipit-source-id: d1bec907485a58f2a3d5532edd631f3f15f73a39