introducing BridgeSynthInlinePass (replacing Bridge- and SynthPass)
Summary:
The BridgePass and SynthPass are old, performing brittle pattern matching that has even most recently caused us problems (see e.g. D30167547).

While they essential implement specialized forms of inlining, implemented at a time before Redex has a powerful general-purpose inliner, they do not check for all the potential gotchas that the inliner does.

For future better-engineering projects such as respecting Java's static initialization rules, it is cumbersome to have to understand and properly fix up these legacy passes, even though they essential just do inlining.

In this diff, the legacy BridgePass and SynthPass are effectively getting replaced by a new variation of our existing inliner, the `BridgeSynthInlinePass`. It collects a set of candidates that is the combination of what the BridgePass and SynthPass were considering, and then it leaves it up to the general purpose inliner to decide whether actually inlining a candidate is beneficial in terms of size, and doesn't violate any other obscure constraints. Note that the selection of candidates here is rather a heuristic for the inliner, mimicking for worked for the BridgePass and SynthPass, but the particular choices do not affect correctness --- the existing inliner takes care of that.

Putting it all together, the main motivation of this diff is better engineering, not achieving additional wins.

Going forward, we can tweak the selection of bridge/synth methods to inline more where beneficial, e.g. boxing/unboxing bridge methods.

Reviewed By: beicy

Differential Revision: D30213491

fbshipit-source-id: e06e0436aba1174e4ba6173ae1850425fa9e56c0