Start refactoring DexTypeList access (1/2)
Summary:
The `Deque` in `DexTypeList` is memory-wise very suboptimal, as common STL implementations use large chunks (512B to 4096B), while the lists are typically short. It is unclear that the optimization of being able to prepend easily is helpful (and inspection of some sites shows that it could just as well be implemented as append).

Start making the caller less aware of the internals where that is easily possible.

Introduce some aliases to the container type and make `DexTypeList` look more like an STL container.

Reviewed By: NTillmann

Differential Revision: D31490792

fbshipit-source-id: b922e081b236c2a7916a10200d89599b35eddb1b