Remove `DexTypeList::get_type_list_internal`
Summary:
Reimplement functionality to remove, prepend, append or replace elements. Move this into `DexTypeList` as a central place.

Using a `std::deque` was misinformed. While the front-insert is `O(1)`, this does not matter as a copy is made of the deque anyways (it is not copy-on-write). At that point, most any container can really be used. The implementation in `DexTypeList` is generic (and relies on a `begin()...end()` copy being as optimized as a direct `const deque&` copy). For that, changing the implementation to a space-conserving `std::vector` should now be mostly drop-in (with the caveat that a `deque` has space preallocated).

Reviewed By: NTillmann

Differential Revision: D31492508

fbshipit-source-id: 4a004eee0d134e1c6dc570a68e9f349ec78d5e85